requirements.txt
# Runtime
msgpack>=1.0.0

# Testing
aiohttp>=3.0.0
pytest>=9.0.0
pytest-asyncio>=1.0.0
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=7.0.0
aioresponses>=0.7.0

# Linting
black>=24.0.0
